requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
djangorestframework==3.12.4
idna==3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
